requirements.txt
aiofiles==24.1.0
fastapi==0.116.1
PyPDF2==3.0.1
python-multipart==0.0.20
uvicorn==0.35.0
tiktoken==0.9.0
streamlit==1.46.1
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.23.0
openai==0.28
python-dotenv==1.1.1
PyMuPDF==1.26.3